Complaint delete password

+++ happyguest-api/app/Http/Controllers/ComplaintController.php
@@ -7,6 +7,9 @@ use App\Http\Resources\ComplaintResource;
 use App\Models\Complaint;
 use App\Models\ComplaintFile;
 use Illuminate\Http\Request;
+use App\Http\Requests\DeleteRequest;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 
 class ComplaintController extends Controller
 {
@@ -163,11 +166,23 @@ class ComplaintController extends Controller
     /**
      * Remove the specified complaint from storage.
      *
+     * @param  DeleteRequest  $request
      * @param  int  $id
      * @return JsonResponse
      */
-    public function destroy(int $id)
+    public function destroy(DeleteRequest $request, int $id)
     {
+        // Check if password is correct
+        if (!Hash::check($request->password, Auth::user()->password)) {
+            return response()->json([
+                'errors' => [
+                    'password' => [
+                        __('auth.password'),
+                    ],
+                ],
+            ], 401);
+        }
+
         Complaint::findOrFail($id)->delete();
 
         return response()->json([
